+++ lib/ruta.dart
@@ -3,14 +3,17 @@ library;
 
 export 'package:ruta/src/core/_internal.dart'
     show
+        CorsHeadersMiddleware,
+        HandlerUse,
+        LogRequestsMiddleware,
+        Middleware,
+        MiddlewareFunc,
         Request,
         Response,
         Router,
         RouterEntry,
-        corsHeaders,
         fromShelfHandler,
         fromShelfMiddleware,
-        logRequests,
         serve,
         toShelfHandler,
         toShelfMiddleware;
@@ -20,7 +23,6 @@ export 'package:ruta/src/core/endpoint.dart';
 export 'package:ruta/src/core/handlers.dart';
 export 'package:ruta/src/core/hot_reload.dart';
 export 'package:ruta/src/core/http_method.dart';
-export 'package:ruta/src/core/middleware.dart';
 export 'package:ruta/src/core/pipeline.dart';
 export 'package:ruta/src/core/route.dart';
 export 'package:ruta/src/core/validation/field.dart';