+++ lib/src/core/pipeline.dart
@@ -1,4 +1,5 @@
 import 'package:ruta/ruta.dart';
+import 'package:ruta/src/core/_internal.dart';
 
 /// A helper that makes it easy to compose a set of [Middleware] and a
 /// [Handler].
@@ -12,7 +13,7 @@ class Pipeline {
   /// [middleware] will be the last [Middleware] to process a request and
   /// the first to process a response.
   Pipeline addMiddleware(Middleware middleware) =>
-      _Pipeline(middleware, addHandler);
+      _Pipeline(middleware.call, addHandler);
 
   /// Returns a new [Handler] with [handler] as the final processor of a
   /// [Request] if all of the middleware in the pipeline have passed the request
@@ -23,8 +24,8 @@ class Pipeline {
 class _Pipeline extends Pipeline {
   _Pipeline(this._middleware, this._parent);
 
-  final Middleware _middleware;
-  final Middleware _parent;
+  final MiddlewareFunc _middleware;
+  final MiddlewareFunc _parent;
 
   @override
   Handler addHandler(Handler handler) => _parent(_middleware(handler));